Move child.js to workers folder
